+++ subtasks.ts
@@ -1,9 +1,6 @@
 /* eslint-disable no-await-in-loop, import/no-extraneous-dependencies */
-import {
-  DeploymentsExtension,
-  DeployOptionsBase,
-} from 'hardhat-deploy/dist/types';
-import { fromWei, toChecksumAddress, toWei, numberToHex } from 'web3-utils';
+import { DeploymentsExtension } from 'hardhat-deploy/dist/types';
+import { fromWei, numberToHex, toChecksumAddress, toWei } from 'web3-utils';
 
 import {
   deleteJob,
@@ -20,7 +17,6 @@ import {
   StackticalConfiguration,
 } from './types';
 import {
-  DSLA__factory,
   ERC20__factory,
   IMessenger__factory,
   MessengerRegistry__factory,
@@ -33,7 +29,13 @@ import {
   SLARegistry__factory,
   StakeRegistry__factory,
 } from './typechain';
-import { CONTRACT_NAMES, PERIOD_STATUS, PERIOD_TYPE } from './constants';
+
+import {
+  CONTRACT_NAMES,
+  PERIOD_STATUS,
+  PERIOD_TYPE,
+  TOKEN_NAMES,
+} from './constants';
 import {
   bootstrapStrings,
   generateBootstrapPeriods,
@@ -44,14 +46,12 @@ import {
 import axios from 'axios';
 
 const prettier = require('prettier');
-const { DataFile } = require('edit-config');
 const appRoot = require('app-root-path');
 const fs = require('fs');
 const path = require('path');
 const yaml = require('js-yaml');
 const compose = require('docker-compose');
 const moment = require('moment');
-const bs58 = require('bs58');
 
 export enum SUB_TASK_NAMES {
   PREPARE_CHAINLINK_NODES = 'PREPARE_CHAINLINK_NODES',
@@ -73,6 +73,7 @@ export enum SUB_TASK_NAMES {
   BOOTSTRAP_STAKE_REGISTRY = 'BOOTSTRAP_STAKE_REGISTRY',
   SET_CONTRACTS_ALLOWANCE = 'SET_CONTRACTS_ALLOWANCE',
   REQUEST_SLI = 'REQUEST_SLI',
+  RETRY_REQUEST_SLI = 'RETRY_REQUEST_SLI',
   GET_PRECOORDINATOR = 'GET_PRECOORDINATOR',
   SET_PRECOORDINATOR = 'SET_PRECOORDINATOR',
   DEPLOY_LOCAL_CHAINLINK_NODES = 'DEPLOY_LOCAL_CHAINLINK_NODES',
@@ -464,6 +465,14 @@ subtask(
       abi: artifact.abi,
     });
   }
+  for (let token of stacktical.tokens) {
+    if (token.address) {
+      await deployments.save(token.name, {
+        address: token.address,
+        abi: token.factory.abi,
+      });
+    }
+  }
 });
 
 subtask(SUB_TASK_NAMES.SAVE_CONTRACTS_ADDRESSES, undefined).setAction(
@@ -630,11 +639,11 @@ subtask(SUB_TASK_NAMES.EXPORT_ABIS, undefined).setAction(
 subtask(SUB_TASK_NAMES.BOOTSTRAP_STAKE_REGISTRY, undefined).setAction(
   async (_, hre: any) => {
     const {
-      stacktical: { bootstrap },
+      stacktical: { bootstrap, tokens },
     }: { stacktical: StackticalConfiguration } = hre.network.config;
     const {
       registry: {
-        stake: { allowedTokens, stakingParameters },
+        stake: { stakingParameters },
       },
     } = bootstrap;
     const { deployments, ethers, getNamedAccounts, network } = hre;
@@ -654,18 +663,18 @@ subtask(SUB_TASK_NAMES.BOOTSTRAP_STAKE_REGISTRY, undefined).setAction(
     );
 
     console.log('Allowing tokens on StakeRegistry');
-    for (let tokenName of allowedTokens) {
-      console.log('Allowing ' + tokenName + ' token');
-      const tokenArtifact = await get(tokenName);
+    for (let token of tokens) {
+      console.log('Allowing ' + token.name + ' token');
+      const tokenArtifact = await get(token.name);
       const allowedToken = await stakeRegistry.isAllowedToken(
         tokenArtifact.address
       );
       if (allowedToken) {
-        console.log(tokenName + ' token already allowed');
+        console.log(token.name + ' token already allowed');
       } else {
         const tx = await stakeRegistry.addAllowedTokens(tokenArtifact.address);
         await tx.wait();
-        console.log(tokenName + ' token successfully allowed');
+        console.log(token.name + ' token successfully allowed');
       }
     }
 
@@ -724,6 +733,9 @@ subtask(SUB_TASK_NAMES.BOOTSTRAP_STAKE_REGISTRY, undefined).setAction(
           stakingParameters.maxTokenLength ||
             currentStakingParameters.maxTokenLength,
           stakingParameters.maxLeverage || currentStakingParameters.maxLeverage,
+          stakingParameters.burnDSLA !== undefined
+            ? stakingParameters.burnDSLA
+            : currentStakingParameters.burnDSLA,
           {
             ...(network.config.gas !== 'auto' && {
               gasLimit: network.config.gas,
@@ -756,6 +768,7 @@ subtask(SUB_TASK_NAMES.BOOTSTRAP_STAKE_REGISTRY, undefined).setAction(
           'maxTokenLength: ' + newParameters.maxTokenLength.toString()
         );
         console.log('maxLeverage: ' + newParameters.maxLeverage.toString());
+        console.log('burnDSLA: ' + newParameters.burnDSLA);
       }
     }
     console.log(finishBootstrap);
@@ -990,6 +1003,8 @@ subtask(SUB_TASK_NAMES.DEPLOY_SLA, undefined).setAction(async (_, hre: any) => {
   const { deployer, notDeployer } = await getNamedAccounts();
   const signer = await ethers.getSigner(deployer);
   const { get } = deployments;
+  const { stacktical }: { stacktical: StackticalConfiguration } =
+    hre.network.config;
   console.log('Starting SLA deployment process');
   const { deploy_sla }: { deploy_sla: DeploySLAConfiguration } = scripts;
   const {
@@ -1018,7 +1033,10 @@ subtask(SUB_TASK_NAMES.DEPLOY_SLA, undefined).setAction(async (_, hre: any) => {
   const stakeRegistryArtifact = await get(CONTRACT_NAMES.StakeRegistry);
   const dslaTokenArtifact = await get(CONTRACT_NAMES.DSLA);
   const slaRegistryArtifact = await get(CONTRACT_NAMES.SLARegistry);
-  const dslaToken = await DSLA__factory.connect(
+  const dslaTokenConfig = stacktical.tokens.find(
+    (token) => token.name === TOKEN_NAMES.DSLA
+  );
+  const dslaToken = await dslaTokenConfig.factory.connect(
     dslaTokenArtifact.address,
     signer
   );
@@ -1883,5 +1901,51 @@ subtask(SUB_TASK_NAMES.GET_REVERT_MESSAGE, undefined).setAction(
     });
     console.log('Error data:');
     console.log(data);
+  }
+);
+
+subtask(SUB_TASK_NAMES.RETRY_REQUEST_SLI, undefined).setAction(
+  async (taskArgs, hre: any) => {
+    const { deployments, ethers, getNamedAccounts, network } = hre;
+    const { stacktical }: { stacktical: StackticalConfiguration } =
+      network.config;
+    const { get } = deployments;
+    const { deployer } = await getNamedAccounts();
+    const slaRegistry = await SLARegistry__factory.connect(
+      (
+        await get(CONTRACT_NAMES.SLARegistry)
+      ).address,
+      await ethers.getSigner(deployer)
+    );
+    const slaAddress = taskArgs.slaAddress
+      ? ethers.utils.getAddress(taskArgs.slaAddress)
+      : (await slaRegistry.userSLAs(deployer)).slice(-1)[0];
+
+    console.log('Retrying request:');
+    const messengerContractName = stacktical.bootstrap.registry.messengers.find(
+      (messenger) => messenger.useCaseName === taskArgs.useCaseName
+    ).contract;
+    const messenger = IMessenger__factory.connect(
+      (await get(messengerContractName)).address,
+      await ethers.getSigner(deployer)
+    );
+
+    await messenger.retryRequest(slaAddress, taskArgs.periodId);
+
+    const sla = SLA__factory.connect(
+      slaAddress,
+      await ethers.getSigner(deployer)
+    );
+    await new Promise((resolve) => {
+      sla.on('SLICreated', () => {
+        resolve(null);
+      });
+    });
+    const createdSLI = await sla.periodSLIs(taskArgs.periodId);
+    const { timestamp, sli, status } = createdSLI;
+    console.log('Created SLI timestamp: ', timestamp.toString());
+    console.log('Created SLI sli: ', sli.toString());
+    console.log('Created SLI status: ', PERIOD_STATUS[status]);
+    console.log('SLI request process finished');
   }
 );